Final ver. MatrixDefault.java + some tests for Determinant

+++ src/main/java/interfaces/IMatrix.java
@@ -1,7 +1,7 @@
 package interfaces;
 
 public interface IMatrix {
-    public int getElement(int row, int col);
-    public void setElement(int row, int col, int value);
+    public double getElement(int row, int col);
+    public void setElement(int row, int col, double value);
     public double getDeterminant();
 }
